Move Enumerable, WithName and UserAnswers to models package
