+++ src/main/g8/app/forms/mappings/Formatters.scala
@@ -2,7 +2,7 @@ package forms.mappings
 
 import play.api.data.FormError
 import play.api.data.format.Formatter
-import utils.Enumerable
+import models.Enumerable
 
 import scala.util.control.Exception.nonFatalCatch
