+++ src/main/g8/app/models/WithName.scala
@@ -1,4 +1,4 @@
-package utils
+package models
 
 class WithName(string: String) {
   override val toString: String = string
