+++ src/main/g8/app/models/requests/DataRequest.scala
@@ -1,7 +1,7 @@
 package models.requests
 
 import play.api.mvc.{Request, WrappedRequest}
-import utils.UserAnswers
+import models.UserAnswers
 
 case class OptionalDataRequest[A] (request: Request[A], internalId: String, userAnswers: Option[UserAnswers]) extends WrappedRequest[A](request)
