+++ src/main/g8/app/navigation/Navigator.scala
@@ -5,8 +5,7 @@ import javax.inject.{Inject, Singleton}
 import play.api.mvc.Call
 import controllers.routes
 import pages._
-import models.{CheckMode, Mode, NormalMode}
-import utils.UserAnswers
+import models._
 
 @Singleton
 class Navigator @Inject()() {
