+++ src/main/g8/app/utils/CheckYourAnswersHelper.scala
@@ -1,7 +1,7 @@
 package utils
 
 import controllers.routes
-import models.CheckMode
+import models.{CheckMode, UserAnswers}
 import pages._
 import viewmodels.{AnswerRow, RepeaterAnswerRow, RepeaterAnswerSection}
