+++ src/main/g8/test/models/WithNameSpec.scala
@@ -1,4 +1,4 @@
-package utils
+package models
 
 import org.scalatest.{MustMatchers, WordSpec}
