+++ src/main/g8/test/navigation/FakeNavigator.scala
@@ -2,8 +2,7 @@ package navigation
 
 import play.api.mvc.Call
 import pages._
-import models.{Mode, NormalMode}
-import utils.UserAnswers
+import models.{Mode, NormalMode, UserAnswers}
 
 class FakeNavigator(desiredRoute: Call, mode: Mode = NormalMode) extends Navigator {
   override def nextPage(page: Page, mode: Mode): UserAnswers => Call = _ => desiredRoute
